Removes the (intermediate/experimental) exports and comes with some cleanup

+++ Tools/obsidian_to_mkdocs.sh
@@ -4,18 +4,20 @@
 # https://github.com/Jackiexiao/mkdocs-roamlinks-plugin
 # https://github.com/lukasgeiter/mkdocs-awesome-pages-plugin
 
-# source venv/bin/activate
+source venv/bin/activate
 
-# if [[ -d docs ]]; then
-#     echo "outch"
-#     exit 10
-# fi
+if [[ ! -d Wiki ]]; then
+    echo "outch"
+    exit 10
+fi
 
+source venv/bin/activate
+
+rm -rf site
 rm -rf docs
 cp -rp Wiki docs
-#mv docs/README.md docs/index.md
-#python Tools/obsidian_to_mkdocs.py
+
+python Tools/obsidian_to_mkdocs.py
 
 #mkdocs build
 #mkdocs serve
-
